pyproject.toml
[project]
name = "llm-batch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.55.0",
    "cyclopts>=3.14.0",
    "ipykernel>=6.29.5",
    "litellm>=1.73.6",
    "openai>=1.76.0",
    "polars>=1.28.1",
    "pymupdf>=1.25.5",
    "python-dotenv[cli]>=1.1.0",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.9.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.4.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
llm-batch = "llm_batch.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_batch"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
]
